+++ banner/src/main/java/com/youth/banner/Banner.java
@@ -3,153 +3,59 @@ package com.youth.banner;
 import android.content.Context;
 import android.content.res.TypedArray;
 import android.os.Build;
-import android.support.v4.view.PagerAdapter;
+import android.support.annotation.NonNull;
+import android.support.annotation.Nullable;
+import android.support.annotation.RequiresApi;
 import android.support.v4.view.ViewPager;
 import android.util.AttributeSet;
-import android.util.DisplayMetrics;
 import android.util.Log;
-import android.util.TypedValue;
-import android.view.Gravity;
-import android.view.LayoutInflater;
-import android.view.MotionEvent;
 import android.view.View;
-import android.view.ViewGroup;
-import android.widget.FrameLayout;
 import android.widget.ImageView;
-import android.widget.ImageView.ScaleType;
-import android.widget.LinearLayout;
-import android.widget.RelativeLayout;
-import android.widget.TextView;
 
 import com.youth.banner.listener.OnBannerClickListener;
-import com.youth.banner.listener.OnBannerListener;
 import com.youth.banner.loader.ImageLoaderInterface;
-import com.youth.banner.view.BannerViewPager;
 
-import java.lang.reflect.Field;
-import java.util.ArrayList;
 import java.util.List;
 
-import static android.support.v4.view.ViewPager.OnPageChangeListener;
-import static android.support.v4.view.ViewPager.PageTransformer;
 
-public class Banner extends FrameLayout implements OnPageChangeListener {
-    public String tag = "banner";
-    private int mIndicatorMargin = BannerConfig.PADDING_SIZE;
-    private int mIndicatorWidth;
-    private int mIndicatorHeight;
-    private int indicatorSize;
-    private int bannerBackgroundImage;
-    private int bannerStyle = BannerConfig.CIRCLE_INDICATOR;
-    private int delayTime = BannerConfig.TIME;
-    private int scrollTime = BannerConfig.DURATION;
-    private boolean isAutoPlay = BannerConfig.IS_AUTO_PLAY;
-    private boolean isScroll = BannerConfig.IS_SCROLL;
-    private int mIndicatorSelectedResId = R.drawable.gray_radius;
-    private int mIndicatorUnselectedResId = R.drawable.white_radius;
-    private int mLayoutResId = R.layout.banner;
-    private int titleHeight;
-    private int titleBackground;
-    private int titleTextColor;
-    private int titleTextSize;
-    private int count = 0;
-    private int currentItem;
-    private int gravity = -1;
-    private int lastPosition = 1;
+public class Banner extends LoopLayout {
+    private ImageLoaderInterface<View> imageLoader;
     private int scaleType = 1;
-    private List<String> titles;
-    private List imageUrls;
-    private List<View> imageViews;
-    private List<ImageView> indicatorImages;
-    private Context context;
-    private BannerViewPager viewPager;
-    private TextView bannerTitle, numIndicatorInside, numIndicator;
-    private LinearLayout indicator, indicatorInside, titleView;
-    private ImageView bannerDefaultImage;
-    private ImageLoaderInterface imageLoader;
-    private BannerPagerAdapter adapter;
-    private OnPageChangeListener mOnPageChangeListener;
-    private BannerScroller mScroller;
-    private OnBannerClickListener bannerListener;
-    private OnBannerListener listener;
-    private DisplayMetrics dm;
-
-    private WeakHandler handler = new WeakHandler();
+    private List<?> imagesUrls;
+    private OnBannerClickListener bannerListener;//only supported in Banner
+    private LoopLayout.Builder builder;
 
     public Banner(Context context) {
-        this(context, null);
+        super(context, null);
     }
 
     public Banner(Context context, AttributeSet attrs) {
-        this(context, attrs, 0);
+        super(context, attrs, 0);
     }
 
     public Banner(Context context, AttributeSet attrs, int defStyle) {
         super(context, attrs, defStyle);
-        this.context = context;
-        titles = new ArrayList<>();
-        imageUrls = new ArrayList<>();
-        imageViews = new ArrayList<>();
-        indicatorImages = new ArrayList<>();
-        dm = context.getResources().getDisplayMetrics();
-        indicatorSize = dm.widthPixels / 80;
-        initView(context, attrs);
     }
 
-    private void initView(Context context, AttributeSet attrs) {
-        imageViews.clear();
-        handleTypedArray(context, attrs);
-        View view = LayoutInflater.from(context).inflate(mLayoutResId, this, true);
-        bannerDefaultImage = (ImageView) view.findViewById(R.id.bannerDefaultImage);
-        viewPager = (BannerViewPager) view.findViewById(R.id.bannerViewPager);
-        titleView = (LinearLayout) view.findViewById(R.id.titleView);
-        indicator = (LinearLayout) view.findViewById(R.id.circleIndicator);
-        indicatorInside = (LinearLayout) view.findViewById(R.id.indicatorInside);
-        bannerTitle = (TextView) view.findViewById(R.id.bannerTitle);
-        numIndicator = (TextView) view.findViewById(R.id.numIndicator);
-        numIndicatorInside = (TextView) view.findViewById(R.id.numIndicatorInside);
-        bannerDefaultImage.setImageResource(bannerBackgroundImage);
-        initViewPagerScroll();
+    @RequiresApi(api = Build.VERSION_CODES.LOLLIPOP)
+    public Banner(@NonNull Context context, @Nullable AttributeSet attrs, int defStyleAttr, int defStyleRes) {
+        super(context, attrs, defStyleAttr, defStyleRes);
     }
 
-    private void handleTypedArray(Context context, AttributeSet attrs) {
-        if (attrs == null) {
-            return;
-        }
-        TypedArray typedArray = context.obtainStyledAttributes(attrs, R.styleable.Banner);
-        mIndicatorWidth = typedArray.getDimensionPixelSize(R.styleable.Banner_indicator_width, indicatorSize);
-        mIndicatorHeight = typedArray.getDimensionPixelSize(R.styleable.Banner_indicator_height, indicatorSize);
-        mIndicatorMargin = typedArray.getDimensionPixelSize(R.styleable.Banner_indicator_margin, BannerConfig.PADDING_SIZE);
-        mIndicatorSelectedResId = typedArray.getResourceId(R.styleable.Banner_indicator_drawable_selected, R.drawable.gray_radius);
-        mIndicatorUnselectedResId = typedArray.getResourceId(R.styleable.Banner_indicator_drawable_unselected, R.drawable.white_radius);
+    @Override
+    protected void init(@NonNull Context context, @Nullable AttributeSet attrs) {
+        super.init(context, attrs);
+        builder = new Builder(this);
+    }
+
+    @Override
+    protected void initStyleable(TypedArray typedArray) {
+        super.initStyleable(typedArray);
         scaleType = typedArray.getInt(R.styleable.Banner_image_scale_type, scaleType);
-        delayTime = typedArray.getInt(R.styleable.Banner_delay_time, BannerConfig.TIME);
-        scrollTime = typedArray.getInt(R.styleable.Banner_scroll_time, BannerConfig.DURATION);
-        isAutoPlay = typedArray.getBoolean(R.styleable.Banner_is_auto_play, BannerConfig.IS_AUTO_PLAY);
-        titleBackground = typedArray.getColor(R.styleable.Banner_title_background, BannerConfig.TITLE_BACKGROUND);
-        titleHeight = typedArray.getDimensionPixelSize(R.styleable.Banner_title_height, BannerConfig.TITLE_HEIGHT);
-        titleTextColor = typedArray.getColor(R.styleable.Banner_title_textcolor, BannerConfig.TITLE_TEXT_COLOR);
-        titleTextSize = typedArray.getDimensionPixelSize(R.styleable.Banner_title_textsize, BannerConfig.TITLE_TEXT_SIZE);
-        mLayoutResId = typedArray.getResourceId(R.styleable.Banner_banner_layout, mLayoutResId);
-        bannerBackgroundImage = typedArray.getResourceId(R.styleable.Banner_banner_default_image, R.drawable.no_banner);
-        typedArray.recycle();
     }
-
-    private void initViewPagerScroll() {
-        try {
-            Field mField = ViewPager.class.getDeclaredField("mScroller");
-            mField.setAccessible(true);
-            mScroller = new BannerScroller(viewPager.getContext());
-            mScroller.setDuration(scrollTime);
-            mField.set(viewPager, mScroller);
-        } catch (Exception e) {
-            Log.e(tag, e.getMessage());
-        }
-    }
-
 
     public Banner isAutoPlay(boolean isAutoPlay) {
-        this.isAutoPlay = isAutoPlay;
+        builder.isAutoPlay(isAutoPlay);
         return this;
     }
 
@@ -159,31 +65,17 @@ public class Banner extends FrameLayout implements OnPageChangeListener {
     }
 
     public Banner setDelayTime(int delayTime) {
-        this.delayTime = delayTime;
+        builder.setDelayTime(delayTime);
         return this;
     }
 
     public Banner setIndicatorGravity(int type) {
-        switch (type) {
-            case BannerConfig.LEFT:
-                this.gravity = Gravity.LEFT | Gravity.CENTER_VERTICAL;
-                break;
-            case BannerConfig.CENTER:
-                this.gravity = Gravity.CENTER;
-                break;
-            case BannerConfig.RIGHT:
-                this.gravity = Gravity.RIGHT | Gravity.CENTER_VERTICAL;
-                break;
-        }
+        builder.setIndicatorGravity(type);
         return this;
     }
 
-    public Banner setBannerAnimation(Class<? extends PageTransformer> transformer) {
-        try {
-            setPageTransformer(true, transformer.newInstance());
-        } catch (Exception e) {
-            Log.e(tag, "Please set the PageTransformer class");
-        }
+    public LoopLayout setBannerAnimation(Class<? extends ViewPager.PageTransformer> transformer) {
+        builder.setBannerAnimation(transformer);
         return this;
     }
 
@@ -195,15 +87,13 @@ public class Banner extends FrameLayout implements OnPageChangeListener {
      * @param limit How many pages will be kept offscreen in an idle state.
      * @return Banner
      */
-    public Banner setOffscreenPageLimit(int limit) {
-        if (viewPager != null) {
-            viewPager.setOffscreenPageLimit(limit);
-        }
+    public LoopLayout setOffscreenPageLimit(int limit) {
+        builder.setOffscreenPageLimit(limit);
         return this;
     }
 
     /**
-     * Set a {@link PageTransformer} that will be called for each attached page whenever
+     * Set a {@link ViewPager.PageTransformer} that will be called for each attached page whenever
      * the scroll position is changed. This allows the application to apply custom property
      * transformations to each page, overriding the default sliding look and feel.
      *
@@ -212,156 +102,68 @@ public class Banner extends FrameLayout implements OnPageChangeListener {
      * @param transformer         PageTransformer that will modify each page's animation properties
      * @return Banner
      */
-    public Banner setPageTransformer(boolean reverseDrawingOrder, PageTransformer transformer) {
-        viewPager.setPageTransformer(reverseDrawingOrder, transformer);
+    public Banner setPageTransformer(boolean reverseDrawingOrder, ViewPager.PageTransformer transformer) {
+        builder.setPageTransformer(reverseDrawingOrder, transformer);
         return this;
     }
 
     public Banner setBannerTitles(List<String> titles) {
-        this.titles = titles;
+        builder.setBannerTitles(titles);
         return this;
     }
 
     public Banner setBannerStyle(int bannerStyle) {
-        this.bannerStyle = bannerStyle;
+        builder.setBannerStyle(bannerStyle);
         return this;
     }
 
     public Banner setViewPagerIsScroll(boolean isScroll) {
-        this.isScroll = isScroll;
+        builder.setViewPagerIsScroll(isScroll);
         return this;
     }
 
+
+    /**
+     * From now on ,we don't need to know your data source but only the count ,items can be init by overwriting {@link #getItemView(int)}.
+     * So Use {@link #setItemCount(int)} instead.
+     */
     public Banner setImages(List<?> imageUrls) {
-        this.imageUrls = imageUrls;
-        this.count = imageUrls.size();
+        this.imagesUrls = imageUrls;
+        setItemCount(imageUrls == null ? 0 : imageUrls.size());
         return this;
     }
 
+    /**
+     * @see #setImages(List)  For the same reason,use {@link #update(int, List)} instead.
+     */
     public void update(List<?> imageUrls, List<String> titles) {
-        this.titles.clear();
-        this.titles.addAll(titles);
-        update(imageUrls);
+        this.imagesUrls = imageUrls;
+        update(imageUrls == null ? 0 : imageUrls.size(), titles);
     }
 
+    /**
+     * @see #setImages(List) For the same reason, use {@link #update(int)} instead.
+     */
     public void update(List<?> imageUrls) {
-        this.imageUrls.clear();
-        this.imageViews.clear();
-        this.indicatorImages.clear();
-        this.imageUrls.addAll(imageUrls);
-        this.count = this.imageUrls.size();
-        start();
+        this.imagesUrls = imageUrls;
+        update(imageUrls == null ? 0 : imageUrls.size());
     }
 
-    public void updateBannerStyle(int bannerStyle) {
-        indicator.setVisibility(GONE);
-        numIndicator.setVisibility(GONE);
-        numIndicatorInside.setVisibility(GONE);
-        indicatorInside.setVisibility(GONE);
-        bannerTitle.setVisibility(View.GONE);
-        titleView.setVisibility(View.GONE);
-        this.bannerStyle = bannerStyle;
-        start();
-    }
-
-    public Banner start() {
-        setBannerStyleUI();
-        setImageList(imageUrls);
-        setData();
-        return this;
-    }
-
-    private void setTitleStyleUI() {
-        if (titles.size() != imageUrls.size()) {
-            throw new RuntimeException("[Banner] --> The number of titles and images is different");
+    @Override
+    protected View getItemView(int realPosition) {
+        View imageView = null;
+        if (imageLoader != null) {
+            imageView = imageLoader.createImageView(context);
         }
-        if (titleBackground != -1) {
-            titleView.setBackgroundColor(titleBackground);
+        if (imageView == null) {
+            imageView = new ImageView(context);
         }
-        if (titleHeight != -1) {
-            titleView.setLayoutParams(new RelativeLayout.LayoutParams(ViewGroup.LayoutParams.MATCH_PARENT, titleHeight));
-        }
-        if (titleTextColor != -1) {
-            bannerTitle.setTextColor(titleTextColor);
-        }
-        if (titleTextSize != -1) {
-            bannerTitle.setTextSize(TypedValue.COMPLEX_UNIT_PX, titleTextSize);
-        }
-        if (titles != null && titles.size() > 0) {
-            bannerTitle.setText(titles.get(0));
-            bannerTitle.setVisibility(View.VISIBLE);
-            titleView.setVisibility(View.VISIBLE);
-        }
-    }
-
-    private void setBannerStyleUI() {
-        int visibility =count > 1 ? View.VISIBLE :View.GONE;
-        switch (bannerStyle) {
-            case BannerConfig.CIRCLE_INDICATOR:
-                indicator.setVisibility(visibility);
-                break;
-            case BannerConfig.NUM_INDICATOR:
-                numIndicator.setVisibility(visibility);
-                break;
-            case BannerConfig.NUM_INDICATOR_TITLE:
-                numIndicatorInside.setVisibility(visibility);
-                setTitleStyleUI();
-                break;
-            case BannerConfig.CIRCLE_INDICATOR_TITLE:
-                indicator.setVisibility(visibility);
-                setTitleStyleUI();
-                break;
-            case BannerConfig.CIRCLE_INDICATOR_TITLE_INSIDE:
-                indicatorInside.setVisibility(visibility);
-                setTitleStyleUI();
-                break;
-        }
-    }
-
-    private void initImages() {
-        imageViews.clear();
-        if (bannerStyle == BannerConfig.CIRCLE_INDICATOR ||
-                bannerStyle == BannerConfig.CIRCLE_INDICATOR_TITLE ||
-                bannerStyle == BannerConfig.CIRCLE_INDICATOR_TITLE_INSIDE) {
-            createIndicator();
-        } else if (bannerStyle == BannerConfig.NUM_INDICATOR_TITLE) {
-            numIndicatorInside.setText("1/" + count);
-        } else if (bannerStyle == BannerConfig.NUM_INDICATOR) {
-            numIndicator.setText("1/" + count);
-        }
-    }
-
-    private void setImageList(List<?> imagesUrl) {
-        if (imagesUrl == null || imagesUrl.size() <= 0) {
-            bannerDefaultImage.setVisibility(VISIBLE);
-            Log.e(tag, "The image data set is empty.");
-            return;
-        }
-        bannerDefaultImage.setVisibility(GONE);
-        initImages();
-        for (int i = 0; i <= count + 1; i++) {
-            View imageView = null;
-            if (imageLoader != null) {
-                imageView = imageLoader.createImageView(context);
-            }
-            if (imageView == null) {
-                imageView = new ImageView(context);
-            }
-            setScaleType(imageView);
-            Object url = null;
-            if (i == 0) {
-                url = imagesUrl.get(count - 1);
-            } else if (i == count + 1) {
-                url = imagesUrl.get(0);
-            } else {
-                url = imagesUrl.get(i - 1);
-            }
-            imageViews.add(imageView);
-            if (imageLoader != null)
-                imageLoader.displayImage(context, url, imageView);
-            else
-                Log.e(tag, "Please set images loader.");
-        }
+        setScaleType(imageView);
+        if (imageLoader != null)
+            imageLoader.displayImage(context, imagesUrls.get(realPosition), imageView);
+        else
+            Log.e(tag, "Please set images loader.");
+        return imageView;
     }
 
     private void setScaleType(View imageView) {
@@ -369,244 +171,47 @@ public class Banner extends FrameLayout implements OnPageChangeListener {
             ImageView view = ((ImageView) imageView);
             switch (scaleType) {
                 case 0:
-                    view.setScaleType(ScaleType.CENTER);
+                    view.setScaleType(ImageView.ScaleType.CENTER);
                     break;
                 case 1:
-                    view.setScaleType(ScaleType.CENTER_CROP);
+                    view.setScaleType(ImageView.ScaleType.CENTER_CROP);
                     break;
                 case 2:
-                    view.setScaleType(ScaleType.CENTER_INSIDE);
+                    view.setScaleType(ImageView.ScaleType.CENTER_INSIDE);
                     break;
                 case 3:
-                    view.setScaleType(ScaleType.FIT_CENTER);
+                    view.setScaleType(ImageView.ScaleType.FIT_CENTER);
                     break;
                 case 4:
-                    view.setScaleType(ScaleType.FIT_END);
+                    view.setScaleType(ImageView.ScaleType.FIT_END);
                     break;
                 case 5:
-                    view.setScaleType(ScaleType.FIT_START);
+                    view.setScaleType(ImageView.ScaleType.FIT_START);
                     break;
                 case 6:
-                    view.setScaleType(ScaleType.FIT_XY);
+                    view.setScaleType(ImageView.ScaleType.FIT_XY);
                     break;
                 case 7:
-                    view.setScaleType(ScaleType.MATRIX);
+                    view.setScaleType(ImageView.ScaleType.MATRIX);
                     break;
             }
 
         }
     }
 
-    private void createIndicator() {
-        indicatorImages.clear();
-        indicator.removeAllViews();
-        indicatorInside.removeAllViews();
-        for (int i = 0; i < count; i++) {
-            ImageView imageView = new ImageView(context);
-            imageView.setScaleType(ScaleType.CENTER_CROP);
-            LinearLayout.LayoutParams params = new LinearLayout.LayoutParams(mIndicatorWidth, mIndicatorHeight);
-            params.leftMargin = mIndicatorMargin;
-            params.rightMargin = mIndicatorMargin;
-            if (i == 0) {
-                imageView.setImageResource(mIndicatorSelectedResId);
-            } else {
-                imageView.setImageResource(mIndicatorUnselectedResId);
-            }
-            indicatorImages.add(imageView);
-            if (bannerStyle == BannerConfig.CIRCLE_INDICATOR ||
-                    bannerStyle == BannerConfig.CIRCLE_INDICATOR_TITLE)
-                indicator.addView(imageView, params);
-            else if (bannerStyle == BannerConfig.CIRCLE_INDICATOR_TITLE_INSIDE)
-                indicatorInside.addView(imageView, params);
-        }
-    }
-
-
-    private void setData() {
-        currentItem = 1;
-        if (adapter == null) {
-            adapter = new BannerPagerAdapter();
-            viewPager.addOnPageChangeListener(this);
-        }
-        viewPager.setAdapter(adapter);
-        viewPager.setFocusable(true);
-        viewPager.setCurrentItem(1);
-        if (gravity != -1)
-            indicator.setGravity(gravity);
-        if (isScroll && count > 1) {
-            viewPager.setScrollable(true);
-        } else {
-            viewPager.setScrollable(false);
-        }
-        if (isAutoPlay)
-            startAutoPlay();
-    }
-
-
-    public void startAutoPlay() {
-        handler.removeCallbacks(task);
-        handler.postDelayed(task, delayTime);
-    }
-
-    public void stopAutoPlay() {
-        handler.removeCallbacks(task);
-    }
-
-    private final Runnable task = new Runnable() {
-        @Override
-        public void run() {
-            if (count > 1 && isAutoPlay) {
-                currentItem = currentItem % (count + 1) + 1;
-//                Log.i(tag, "curr:" + currentItem + " count:" + count);
-                if (currentItem == 1) {
-                    viewPager.setCurrentItem(currentItem, false);
-                    handler.post(task);
-                } else {
-                    viewPager.setCurrentItem(currentItem);
-                    handler.postDelayed(task, delayTime);
+    @Override
+    protected void initItem(View view, final int position) {
+        super.initItem(view, position);
+        if (bannerListener != null) {
+            view.setOnClickListener(new OnClickListener() {
+                @Override
+                public void onClick(View v) {
+                    Log.e(tag, "你正在使用旧版点击事件接口，下标是从1开始，" +
+                            "为了体验请更换为setOnBannerListener，下标从0开始计算");
+                    bannerListener.OnBannerClick(position);
                 }
-            }
+            });
         }
-    };
-
-    @Override
-    public boolean dispatchTouchEvent(MotionEvent ev) {
-//        Log.i(tag, ev.getAction() + "--" + isAutoPlay);
-        if (isAutoPlay) {
-            int action = ev.getAction();
-            if (action == MotionEvent.ACTION_UP || action == MotionEvent.ACTION_CANCEL
-                    || action == MotionEvent.ACTION_OUTSIDE) {
-                startAutoPlay();
-            } else if (action == MotionEvent.ACTION_DOWN) {
-                stopAutoPlay();
-            }
-        }
-        return super.dispatchTouchEvent(ev);
-    }
-
-    /**
-     * 返回真实的位置
-     *
-     * @param position
-     * @return 下标从0开始
-     */
-    public int toRealPosition(int position) {
-        int realPosition = (position - 1) % count;
-        if (realPosition < 0)
-            realPosition += count;
-        return realPosition;
-    }
-
-    class BannerPagerAdapter extends PagerAdapter {
-
-        @Override
-        public int getCount() {
-            return imageViews.size();
-        }
-
-        @Override
-        public boolean isViewFromObject(View view, Object object) {
-            return view == object;
-        }
-
-        @Override
-        public Object instantiateItem(ViewGroup container, final int position) {
-            container.addView(imageViews.get(position));
-            View view = imageViews.get(position);
-            if (bannerListener != null) {
-                view.setOnClickListener(new OnClickListener() {
-                    @Override
-                    public void onClick(View v) {
-                        Log.e(tag, "你正在使用旧版点击事件接口，下标是从1开始，" +
-                                "为了体验请更换为setOnBannerListener，下标从0开始计算");
-                        bannerListener.OnBannerClick(position);
-                    }
-                });
-            }
-            if (listener != null) {
-                view.setOnClickListener(new OnClickListener() {
-                    @Override
-                    public void onClick(View v) {
-                        listener.OnBannerClick(toRealPosition(position));
-                    }
-                });
-            }
-            return view;
-        }
-
-        @Override
-        public void destroyItem(ViewGroup container, int position, Object object) {
-            container.removeView((View) object);
-        }
-
-    }
-
-    @Override
-    public void onPageScrollStateChanged(int state) {
-        if (mOnPageChangeListener != null) {
-            mOnPageChangeListener.onPageScrollStateChanged(state);
-        }
-//        Log.i(tag,"currentItem: "+currentItem);
-        switch (state) {
-            case 0://No operation
-                if (currentItem == 0) {
-                    viewPager.setCurrentItem(count, false);
-                } else if (currentItem == count + 1) {
-                    viewPager.setCurrentItem(1, false);
-                }
-                break;
-            case 1://start Sliding
-                if (currentItem == count + 1) {
-                    viewPager.setCurrentItem(1, false);
-                } else if (currentItem == 0) {
-                    viewPager.setCurrentItem(count, false);
-                }
-                break;
-            case 2://end Sliding
-                break;
-        }
-    }
-
-    @Override
-    public void onPageScrolled(int position, float positionOffset, int positionOffsetPixels) {
-        if (mOnPageChangeListener != null) {
-            mOnPageChangeListener.onPageScrolled(toRealPosition(position), positionOffset, positionOffsetPixels);
-        }
-    }
-
-    @Override
-    public void onPageSelected(int position) {
-        currentItem=position;
-        if (mOnPageChangeListener != null) {
-            mOnPageChangeListener.onPageSelected(toRealPosition(position));
-        }
-        if (bannerStyle == BannerConfig.CIRCLE_INDICATOR ||
-                bannerStyle == BannerConfig.CIRCLE_INDICATOR_TITLE ||
-                bannerStyle == BannerConfig.CIRCLE_INDICATOR_TITLE_INSIDE) {
-            indicatorImages.get((lastPosition - 1 + count) % count).setImageResource(mIndicatorUnselectedResId);
-            indicatorImages.get((position - 1 + count) % count).setImageResource(mIndicatorSelectedResId);
-            lastPosition = position;
-        }
-        if (position == 0) position = count;
-        if (position > count) position = 1;
-        switch (bannerStyle) {
-            case BannerConfig.CIRCLE_INDICATOR:
-                break;
-            case BannerConfig.NUM_INDICATOR:
-                numIndicator.setText(position + "/" + count);
-                break;
-            case BannerConfig.NUM_INDICATOR_TITLE:
-                numIndicatorInside.setText(position + "/" + count);
-                bannerTitle.setText(titles.get(position - 1));
-                break;
-            case BannerConfig.CIRCLE_INDICATOR_TITLE:
-                bannerTitle.setText(titles.get(position - 1));
-                break;
-            case BannerConfig.CIRCLE_INDICATOR_TITLE_INSIDE:
-                bannerTitle.setText(titles.get(position - 1));
-                break;
-        }
-
     }
 
     @Deprecated
@@ -615,22 +220,4 @@ public class Banner extends FrameLayout implements OnPageChangeListener {
         return this;
     }
 
-    /**
-     * 废弃了旧版接口，新版的接口下标是从1开始，同时解决下标越界问题
-     *
-     * @param listener
-     * @return
-     */
-    public Banner setOnBannerListener(OnBannerListener listener) {
-        this.listener = listener;
-        return this;
-    }
-
-    public void setOnPageChangeListener(OnPageChangeListener onPageChangeListener) {
-        mOnPageChangeListener = onPageChangeListener;
-    }
-
-    public void releaseBanner() {
-        handler.removeCallbacksAndMessages(null);
-    }
 }
