+++ banner/src/main/java/com/youth/banner/LoopLayout.java
@@ -1,0 +1,587 @@
+package com.youth.banner;
+
+import android.content.Context;
+import android.content.res.TypedArray;
+import android.os.Build;
+import android.support.annotation.CallSuper;
+import android.support.annotation.NonNull;
+import android.support.annotation.Nullable;
+import android.support.annotation.RequiresApi;
+import android.support.v4.view.PagerAdapter;
+import android.support.v4.view.ViewPager;
+import android.util.AttributeSet;
+import android.util.DisplayMetrics;
+import android.util.Log;
+import android.util.TypedValue;
+import android.view.Gravity;
+import android.view.LayoutInflater;
+import android.view.MotionEvent;
+import android.view.View;
+import android.view.ViewGroup;
+import android.widget.FrameLayout;
+import android.widget.ImageView;
+import android.widget.LinearLayout;
+import android.widget.RelativeLayout;
+import android.widget.TextView;
+
+import com.youth.banner.listener.OnBannerListener;
+import com.youth.banner.view.BannerViewPager;
+
+import java.lang.reflect.Field;
+import java.util.ArrayList;
+import java.util.List;
+
+/**
+ * You can extends this class and custom your own item view.{@link LoopLayout} supports View.If your banner's item is ImageViews,you can still user{@link Banner}like before.
+ */
+public abstract class LoopLayout extends FrameLayout implements ViewPager.OnPageChangeListener {
+    public final String tag = getClass().getSimpleName();
+    private int mIndicatorMargin = BannerConfig.PADDING_SIZE;
+    private int mIndicatorWidth;
+    private int mIndicatorHeight;
+    private int indicatorSize;
+    private int bannerBackgroundImage;
+    private int bannerStyle = BannerConfig.CIRCLE_INDICATOR;
+    private int delayTime = BannerConfig.TIME;
+    private int scrollTime = BannerConfig.DURATION;
+    private boolean isAutoPlay = BannerConfig.IS_AUTO_PLAY;
+    private boolean isScroll = BannerConfig.IS_SCROLL;
+    private int mIndicatorSelectedResId = R.drawable.gray_radius;
+    private int mIndicatorUnselectedResId = R.drawable.white_radius;
+    private int mLayoutResId = R.layout.banner;
+    private int titleHeight;
+    private int titleBackground;
+    private int titleTextColor;
+    private int titleTextSize;
+    private int count = 0;
+    private int currentItem;
+    private int gravity = -1;
+    private int lastPosition = 1;
+    private List<String> titles;
+    private List<View> imageViews;
+    private List<ImageView> indicatorImages;
+    protected Context context;
+    private BannerViewPager viewPager;
+    private TextView bannerTitle, numIndicatorInside, numIndicator;
+    private LinearLayout indicator, indicatorInside, titleView;
+    private ImageView bannerDefaultImage;
+
+    private BannerPagerAdapter adapter;
+    private ViewPager.OnPageChangeListener mOnPageChangeListener;
+    private OnBannerListener listener;
+
+    private WeakHandler handler = new WeakHandler();
+
+    public LoopLayout(@NonNull Context context) {
+        this(context, null);
+    }
+
+    public LoopLayout(@NonNull Context context, @Nullable AttributeSet attrs) {
+        this(context, attrs, 0);
+    }
+
+    public LoopLayout(@NonNull Context context, @Nullable AttributeSet attrs, int defStyleAttr) {
+        super(context, attrs, defStyleAttr);
+        init(context, attrs);
+    }
+
+    @RequiresApi(api = Build.VERSION_CODES.LOLLIPOP)
+    public LoopLayout(@NonNull Context context, @Nullable AttributeSet attrs, int defStyleAttr, int defStyleRes) {
+        super(context, attrs, defStyleAttr, defStyleRes);
+        init(context, attrs);
+    }
+
+    @CallSuper
+    protected void init(@NonNull Context context, @Nullable AttributeSet attrs) {
+        this.context = context;
+        titles = new ArrayList<>();
+        imageViews = new ArrayList<>();
+        indicatorImages = new ArrayList<>();
+        DisplayMetrics dm = context.getResources().getDisplayMetrics();
+        indicatorSize = dm.widthPixels / 80;
+        initView(context, attrs);
+    }
+
+    private void initView(Context context, AttributeSet attrs) {
+        imageViews.clear();
+        handleTypedArray(context, attrs);
+        View view = LayoutInflater.from(context).inflate(mLayoutResId, this, true);
+        bannerDefaultImage = (ImageView) view.findViewById(R.id.bannerDefaultImage);
+        viewPager = (BannerViewPager) view.findViewById(R.id.bannerViewPager);
+        titleView = (LinearLayout) view.findViewById(R.id.titleView);
+        indicator = (LinearLayout) view.findViewById(R.id.circleIndicator);
+        indicatorInside = (LinearLayout) view.findViewById(R.id.indicatorInside);
+        bannerTitle = (TextView) view.findViewById(R.id.bannerTitle);
+        numIndicator = (TextView) view.findViewById(R.id.numIndicator);
+        numIndicatorInside = (TextView) view.findViewById(R.id.numIndicatorInside);
+        bannerDefaultImage.setImageResource(bannerBackgroundImage);
+        initViewPagerScroll();
+    }
+
+    private void handleTypedArray(Context context, AttributeSet attrs) {
+        if (attrs == null) {
+            return;
+        }
+        TypedArray typedArray = context.obtainStyledAttributes(attrs, R.styleable.Banner);
+        initStyleable(typedArray);
+        typedArray.recycle();
+    }
+
+    @CallSuper
+    protected void initStyleable(TypedArray typedArray) {
+        mIndicatorWidth = typedArray.getDimensionPixelSize(R.styleable.Banner_indicator_width, indicatorSize);
+        mIndicatorHeight = typedArray.getDimensionPixelSize(R.styleable.Banner_indicator_height, indicatorSize);
+        mIndicatorMargin = typedArray.getDimensionPixelSize(R.styleable.Banner_indicator_margin, BannerConfig.PADDING_SIZE);
+        mIndicatorSelectedResId = typedArray.getResourceId(R.styleable.Banner_indicator_drawable_selected, R.drawable.gray_radius);
+        mIndicatorUnselectedResId = typedArray.getResourceId(R.styleable.Banner_indicator_drawable_unselected, R.drawable.white_radius);
+        delayTime = typedArray.getInt(R.styleable.Banner_delay_time, BannerConfig.TIME);
+        scrollTime = typedArray.getInt(R.styleable.Banner_scroll_time, BannerConfig.DURATION);
+        isAutoPlay = typedArray.getBoolean(R.styleable.Banner_is_auto_play, BannerConfig.IS_AUTO_PLAY);
+        titleBackground = typedArray.getColor(R.styleable.Banner_title_background, BannerConfig.TITLE_BACKGROUND);
+        titleHeight = typedArray.getDimensionPixelSize(R.styleable.Banner_title_height, BannerConfig.TITLE_HEIGHT);
+        titleTextColor = typedArray.getColor(R.styleable.Banner_title_textcolor, BannerConfig.TITLE_TEXT_COLOR);
+        titleTextSize = typedArray.getDimensionPixelSize(R.styleable.Banner_title_textsize, BannerConfig.TITLE_TEXT_SIZE);
+        mLayoutResId = typedArray.getResourceId(R.styleable.Banner_banner_layout, mLayoutResId);
+        bannerBackgroundImage = typedArray.getResourceId(R.styleable.Banner_banner_default_image, R.drawable.no_banner);
+    }
+
+    private void initViewPagerScroll() {
+        try {
+            Field mField = ViewPager.class.getDeclaredField("mScroller");
+            mField.setAccessible(true);
+            BannerScroller mScroller = new BannerScroller(viewPager.getContext());
+            mScroller.setDuration(scrollTime);
+            mField.set(viewPager, mScroller);
+        } catch (Exception e) {
+            Log.e(tag, e.getMessage());
+        }
+    }
+
+    public LoopLayout setItemCount(int count) {
+        this.count = count;
+        return this;
+    }
+
+    public void update(int count, List<String> titles) {
+        this.titles.clear();
+        this.titles.addAll(titles);
+        update(count);
+    }
+
+    public void update(int count) {
+        this.imageViews.clear();
+        this.indicatorImages.clear();
+        this.count = count;
+        start();
+    }
+
+    public void updateBannerStyle(int bannerStyle) {
+        indicator.setVisibility(GONE);
+        numIndicator.setVisibility(GONE);
+        numIndicatorInside.setVisibility(GONE);
+        indicatorInside.setVisibility(GONE);
+        bannerTitle.setVisibility(View.GONE);
+        titleView.setVisibility(View.GONE);
+        this.bannerStyle = bannerStyle;
+        start();
+    }
+
+    public LoopLayout start() {
+        setBannerStyleUI();
+        setImageList();
+        setData();
+        return this;
+    }
+
+    private void setTitleStyleUI() {
+        if (titles.size() != count) {
+            throw new RuntimeException("[Banner] --> The number of titles and images is different");
+        }
+        if (titleBackground != -1) {
+            titleView.setBackgroundColor(titleBackground);
+        }
+        if (titleHeight != -1) {
+            titleView.setLayoutParams(new RelativeLayout.LayoutParams(ViewGroup.LayoutParams.MATCH_PARENT, titleHeight));
+        }
+        if (titleTextColor != -1) {
+            bannerTitle.setTextColor(titleTextColor);
+        }
+        if (titleTextSize != -1) {
+            bannerTitle.setTextSize(TypedValue.COMPLEX_UNIT_PX, titleTextSize);
+        }
+        if (titles != null && titles.size() > 0) {
+            bannerTitle.setText(titles.get(0));
+            bannerTitle.setVisibility(View.VISIBLE);
+            titleView.setVisibility(View.VISIBLE);
+        }
+    }
+
+    private void setBannerStyleUI() {
+        int visibility = count > 1 ? View.VISIBLE : View.GONE;
+        switch (bannerStyle) {
+            case BannerConfig.CIRCLE_INDICATOR:
+                indicator.setVisibility(visibility);
+                break;
+            case BannerConfig.NUM_INDICATOR:
+                numIndicator.setVisibility(visibility);
+                break;
+            case BannerConfig.NUM_INDICATOR_TITLE:
+                numIndicatorInside.setVisibility(visibility);
+                setTitleStyleUI();
+                break;
+            case BannerConfig.CIRCLE_INDICATOR_TITLE:
+                indicator.setVisibility(visibility);
+                setTitleStyleUI();
+                break;
+            case BannerConfig.CIRCLE_INDICATOR_TITLE_INSIDE:
+                indicatorInside.setVisibility(visibility);
+                setTitleStyleUI();
+                break;
+        }
+    }
+
+    private void initImages() {
+        imageViews.clear();
+        if (bannerStyle == BannerConfig.CIRCLE_INDICATOR ||
+                bannerStyle == BannerConfig.CIRCLE_INDICATOR_TITLE ||
+                bannerStyle == BannerConfig.CIRCLE_INDICATOR_TITLE_INSIDE) {
+            createIndicator();
+        } else if (bannerStyle == BannerConfig.NUM_INDICATOR_TITLE) {
+            numIndicatorInside.setText("1/" + count);
+        } else if (bannerStyle == BannerConfig.NUM_INDICATOR) {
+            numIndicator.setText("1/" + count);
+        }
+    }
+
+    private void setImageList() {
+        if (count <= 0) {
+            bannerDefaultImage.setVisibility(VISIBLE);
+            Log.e(tag, "The data set is empty.");
+            return;
+        }
+        bannerDefaultImage.setVisibility(GONE);
+        initImages();
+        for (int i = 0; i <= count + 1; i++) {
+            imageViews.add(getItemView(toRealPosition(i)));
+        }
+    }
+
+    protected abstract View getItemView(int realPosition);
+
+    private void createIndicator() {
+        indicatorImages.clear();
+        indicator.removeAllViews();
+        indicatorInside.removeAllViews();
+        for (int i = 0; i < count; i++) {
+            ImageView imageView = new ImageView(context);
+            imageView.setScaleType(ImageView.ScaleType.CENTER_CROP);
+            LinearLayout.LayoutParams params = new LinearLayout.LayoutParams(mIndicatorWidth, mIndicatorHeight);
+            params.leftMargin = mIndicatorMargin;
+            params.rightMargin = mIndicatorMargin;
+            if (i == 0) {
+                imageView.setImageResource(mIndicatorSelectedResId);
+            } else {
+                imageView.setImageResource(mIndicatorUnselectedResId);
+            }
+            indicatorImages.add(imageView);
+            if (bannerStyle == BannerConfig.CIRCLE_INDICATOR ||
+                    bannerStyle == BannerConfig.CIRCLE_INDICATOR_TITLE)
+                indicator.addView(imageView, params);
+            else if (bannerStyle == BannerConfig.CIRCLE_INDICATOR_TITLE_INSIDE)
+                indicatorInside.addView(imageView, params);
+        }
+    }
+
+
+    private void setData() {
+        currentItem = 1;
+        if (adapter == null) {
+            adapter = new BannerPagerAdapter();
+            viewPager.addOnPageChangeListener(this);
+        }
+        viewPager.setAdapter(adapter);
+        viewPager.setFocusable(true);
+        viewPager.setCurrentItem(1);
+        if (gravity != -1)
+            indicator.setGravity(gravity);
+        if (isScroll && count > 1) {
+            viewPager.setScrollable(true);
+        } else {
+            viewPager.setScrollable(false);
+        }
+        if (isAutoPlay)
+            startAutoPlay();
+    }
+
+
+    public void startAutoPlay() {
+        handler.removeCallbacks(task);
+        handler.postDelayed(task, delayTime);
+    }
+
+    public void stopAutoPlay() {
+        handler.removeCallbacks(task);
+    }
+
+    private final Runnable task = new Runnable() {
+        @Override
+        public void run() {
+            if (count > 1 && isAutoPlay) {
+                currentItem = currentItem % (count + 1) + 1;
+                if (currentItem == 1) {
+                    viewPager.setCurrentItem(currentItem, false);
+                    handler.post(task);
+                } else {
+                    viewPager.setCurrentItem(currentItem);
+                    handler.postDelayed(task, delayTime);
+                }
+            }
+        }
+    };
+
+    @Override
+    public boolean dispatchTouchEvent(MotionEvent ev) {
+        if (isAutoPlay) {
+            int action = ev.getAction();
+            if (action == MotionEvent.ACTION_UP || action == MotionEvent.ACTION_CANCEL
+                    || action == MotionEvent.ACTION_OUTSIDE) {
+                startAutoPlay();
+            } else if (action == MotionEvent.ACTION_DOWN) {
+                stopAutoPlay();
+            }
+        }
+        return super.dispatchTouchEvent(ev);
+    }
+
+    /**
+     * 返回真实的位置
+     *
+     * @return 下标从0开始
+     */
+    public int toRealPosition(int position) {
+        if (position == 0) {
+            return count - 1;
+        } else if (position == count + 1) {
+            return 0;
+        } else {
+            return position - 1;
+        }
+    }
+
+    class BannerPagerAdapter extends PagerAdapter {
+
+        @Override
+        public int getCount() {
+            return imageViews.size();
+        }
+
+        @Override
+        public boolean isViewFromObject(View view, Object object) {
+            return view == object;
+        }
+
+        @Override
+        public Object instantiateItem(@NonNull ViewGroup container, int position) {
+            container.addView(imageViews.get(position));
+            View view = imageViews.get(position);
+            initItem(view, position);
+            return view;
+        }
+
+        @Override
+        public void destroyItem(ViewGroup container, int position, Object object) {
+            container.removeView((View) object);
+        }
+
+    }
+
+    @CallSuper
+    protected void initItem(View view, final int position) {
+        if (listener != null) {
+            view.setOnClickListener(new OnClickListener() {
+                @Override
+                public void onClick(View v) {
+                    listener.OnBannerClick(toRealPosition(position));
+                }
+            });
+        }
+    }
+
+    @Override
+    public void onPageScrollStateChanged(int state) {
+        if (mOnPageChangeListener != null) {
+            mOnPageChangeListener.onPageScrollStateChanged(state);
+        }
+//        Log.i(tag,"currentItem: "+currentItem);
+        switch (state) {
+            case 0://No operation
+                if (currentItem == 0) {
+                    viewPager.setCurrentItem(count, false);
+                } else if (currentItem == count + 1) {
+                    viewPager.setCurrentItem(1, false);
+                }
+                break;
+            case 1://start Sliding
+                if (currentItem == count + 1) {
+                    viewPager.setCurrentItem(1, false);
+                } else if (currentItem == 0) {
+                    viewPager.setCurrentItem(count, false);
+                }
+                break;
+            case 2://end Sliding
+                break;
+        }
+    }
+
+    @Override
+    public void onPageScrolled(int position, float positionOffset, int positionOffsetPixels) {
+        if (mOnPageChangeListener != null) {
+            mOnPageChangeListener.onPageScrolled(toRealPosition(position), positionOffset, positionOffsetPixels);
+        }
+    }
+
+    @Override
+    public void onPageSelected(int position) {
+        currentItem = position;
+        if (mOnPageChangeListener != null) {
+            mOnPageChangeListener.onPageSelected(toRealPosition(position));
+        }
+        if (bannerStyle == BannerConfig.CIRCLE_INDICATOR ||
+                bannerStyle == BannerConfig.CIRCLE_INDICATOR_TITLE ||
+                bannerStyle == BannerConfig.CIRCLE_INDICATOR_TITLE_INSIDE) {
+            indicatorImages.get((lastPosition - 1 + count) % count).setImageResource(mIndicatorUnselectedResId);
+            indicatorImages.get((position - 1 + count) % count).setImageResource(mIndicatorSelectedResId);
+            lastPosition = position;
+        }
+        if (position == 0) position = count;
+        if (position > count) position = 1;
+        switch (bannerStyle) {
+            case BannerConfig.CIRCLE_INDICATOR:
+                break;
+            case BannerConfig.NUM_INDICATOR:
+                numIndicator.setText(position + "/" + count);
+                break;
+            case BannerConfig.NUM_INDICATOR_TITLE:
+                numIndicatorInside.setText(position + "/" + count);
+                bannerTitle.setText(titles.get(position - 1));
+                break;
+            case BannerConfig.CIRCLE_INDICATOR_TITLE:
+                bannerTitle.setText(titles.get(position - 1));
+                break;
+            case BannerConfig.CIRCLE_INDICATOR_TITLE_INSIDE:
+                bannerTitle.setText(titles.get(position - 1));
+                break;
+        }
+
+    }
+
+
+    /**
+     * A strange builder to be compatible with old code.
+     */
+    public static class Builder {
+        private LoopLayout loopLayout;
+
+        public Builder(LoopLayout loopLayout) {
+            this.loopLayout = loopLayout;
+        }
+
+        public Builder isAutoPlay(boolean isAutoPlay) {
+            loopLayout.isAutoPlay = isAutoPlay;
+            return this;
+        }
+
+        public Builder setBannerTitles(List<String> titles) {
+            loopLayout.titles = titles;
+            return this;
+        }
+
+        public Builder setDelayTime(int delayTime) {
+            loopLayout.delayTime = delayTime;
+            return this;
+        }
+
+        public Builder setIndicatorGravity(int type) {
+            switch (type) {
+                case BannerConfig.LEFT:
+                    loopLayout.gravity = Gravity.LEFT | Gravity.CENTER_VERTICAL;
+                    break;
+                case BannerConfig.CENTER:
+                    loopLayout.gravity = Gravity.CENTER;
+                    break;
+                case BannerConfig.RIGHT:
+                    loopLayout.gravity = Gravity.RIGHT | Gravity.CENTER_VERTICAL;
+                    break;
+            }
+            return this;
+        }
+
+        public Builder setBannerAnimation(Class<? extends ViewPager.PageTransformer> transformer) {
+            try {
+                setPageTransformer(true, transformer.newInstance());
+            } catch (Exception e) {
+                Log.e(loopLayout.tag, "Please set the PageTransformer class");
+            }
+            return this;
+        }
+
+        /**
+         * Set the number of pages that should be retained to either side of the
+         * current page in the view hierarchy in an idle state. Pages beyond this
+         * limit will be recreated from the adapter when needed.
+         *
+         * @param limit How many pages will be kept offscreen in an idle state.
+         */
+        public Builder setOffscreenPageLimit(int limit) {
+            if (loopLayout.viewPager != null) {
+                loopLayout.viewPager.setOffscreenPageLimit(limit);
+            }
+            return this;
+        }
+
+        /**
+         * Set a {@link ViewPager.PageTransformer} that will be called for each attached page whenever
+         * the scroll position is changed. This allows the application to apply custom property
+         * transformations to each page, overriding the default sliding look and feel.
+         *
+         * @param reverseDrawingOrder true if the supplied PageTransformer requires page views
+         *                            to be drawn from last to first instead of first to last.
+         * @param transformer         PageTransformer that will modify each page's animation properties
+         */
+        public Builder setPageTransformer(boolean reverseDrawingOrder, ViewPager.PageTransformer transformer) {
+            loopLayout.viewPager.setPageTransformer(reverseDrawingOrder, transformer);
+            return this;
+        }
+
+        public Builder setBannerStyle(int bannerStyle) {
+            loopLayout.bannerStyle = bannerStyle;
+            return this;
+        }
+
+        public Builder setViewPagerIsScroll(boolean isScroll) {
+            loopLayout.isScroll = isScroll;
+            return this;
+        }
+
+        public void start() {
+            loopLayout.start();
+        }
+    }
+
+
+    /**
+     * 废弃了旧版接口，新版的接口下标是从1开始，同时解决下标越界问题
+     */
+    public LoopLayout setOnBannerListener(OnBannerListener listener) {
+        this.listener = listener;
+        return this;
+    }
+
+    public void setOnPageChangeListener(ViewPager.OnPageChangeListener onPageChangeListener) {
+        mOnPageChangeListener = onPageChangeListener;
+    }
+
+    public void releaseBanner() {
+        handler.removeCallbacksAndMessages(null);
+    }
+}
